Remove decimal places from resolution selection
git-svn-id: https://betaville.googlecode.com/svn/BetavilleApp/trunk@4240 8c047b66-80f2-bdae-70f3-23be71ff3dc6

+++ src/edu/poly/bxmc/betaville/gui/BetavilleSettingsPanel.java
@@ -386,7 +386,7 @@ public class BetavilleSettingsPanel extends JFrame{
 		}
 
 		public String toString(){
-			return dimension.getWidth()+"x"+dimension.getHeight();
+			return (int)dimension.getWidth()+"x"+(int)dimension.getHeight();
 		}
 	}
